Refactor image type code sequence handling in OrthodonticPhotograph; migrate utility function to class method
